+++ ecc/logic/tools.py
@@ -7,7 +7,9 @@ Please credit me if you change, use or adapt this file.
 import copy
 import random
 import shutil
+import os
 import time
+import json
 
 from collections import defaultdict
